+++ PostHog/Surveys/QuestionHeader.swift
@@ -14,7 +14,7 @@
 
         let question: String
         let description: String?
-        let contentType: PostHogSurveyTextContentType
+        let contentType: PostHogDisplaySurveyTextContentType
 
         var body: some View {
             VStack(alignment: .leading, spacing: 8) {
@@ -29,6 +29,7 @@
                         .multilineTextAlignment(.leading)
                 }
             }
+            .padding(.top, 16)
             .frame(maxWidth: .infinity, alignment: .leading)
         }